requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html

torch==1.9.1+cu111
torchvision==0.10.1+cu111

retinaface-pytorch

opencv-python
numpy

streamlit
streamlit-image-comparison
tqdm
seaborn
matplotlib
psutil